+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Development
 
+- sending information event for `device orientation`
+
 ## v1.9.0
 ### Fixed
 - Error message contained too much data, so that grouping is not useful